+++ ui/client/types/process.ts
@@ -29,4 +29,9 @@ export type ProcessDefinitionData = {
   nodesConfig?: $TodoType,
   nodesToAdd?: NodesGroup[],
   processDefinition?: $TodoType,
+  servicesDefinition?: ServicesDefinition
+}
+
+export type ServicesDefinition = {
+  defaultAsyncInterpretation: Boolean
 }