+++ docs/Changelog.md
@@ -10,6 +10,7 @@ Nussknacker versions
 
 0.2.0 (not released yet)
 ------------------------
+* [#1024](https://github.com/TouK/nussknacker/pull/1024) Added default async interpretation value configured by `asyncExecutionConfig.defaultUseAsyncInterpretation` (false if missing).
 * [#879](https://github.com/TouK/nussknacker/pull/879) Metrics can now use Flink variables for better reporting, it's recommended to use InfluxDB native protocol instead of legacy Graphite protocol to send metrics to InfluxDB.
 * [#940](https://github.com/TouK/nussknacker/pull/940) More detailed node errors 
 * [#949](https://github.com/TouK/nussknacker/pull/949) JVM options can be configured via JDK_JAVA_OPTIONS env variable (in docker and standalone distribution)